Anpassung Jenksinsfile

+++ Jenkinsfile
@@ -4,6 +4,10 @@ node {
     stage('Checkout') {
       checkout scm
     }
+    environment {
+      registry = "martinhug/repository_name"
+      registryCredential = 'dockerhub'
+    }
     stage('Environment') {
       sh 'git --version'
       echo "Branch: ${env.BRANCH_NAME}"
@@ -13,14 +17,12 @@ node {
       sh 'printenv'
     }
     stage('Deploy Docker Frontend'){
-
+    if(env.BRANCH_NAME == 'master'){
         sh 'docker build -f frontend/Dockerfile  -t topology-designer-frontend --no-cache .'
         sh 'docker tag topology-designer-frontend localhost:5000/topology-designer-frontend'
         sh 'docker push localhost:5000/topology-designer-frontend'
-        sh 'docker commit topology-designer-frontend martinhug/topology-designer-frontend:latest'
-        sh 'docker push martinhug/topology-designer-frontend:latest'
         sh 'docker rmi -f topology-designer-frontend localhost:5000/topology-designer-frontend'
-
+    }
     }
 
     stage('Deploy Docker Backend'){
